+++ src/main/java/com/example/learnacademy/repository/CourseContentRepository.java
@@ -3,7 +3,8 @@ package com.example.learnacademy.repository;
 import com.example.learnacademy.model.CourseContent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseContentRepository extends JpaRepository<CourseContent, Integer> {
+public interface CourseContentRepository extends JpaRepository<CourseContent, Long> {
 
-    CourseContent findByCourseId(Integer courseId);
+    CourseContent findByCourseId(Long courseId);
+
 }
