+++ src/main/java/com/example/learnacademy/repository/UserCourseRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserCourseRepository extends JpaRepository<UserCourse, Integer> {
+public interface UserCourseRepository extends JpaRepository<UserCourse, Long> {
 
-    List<UserCourse> findByUserId(Integer userId);
+    List<UserCourse> findByUserId(Long userId);
 
 }
